refactor: consolidate 8 wrapper pages into single files
Merge routing/data-fetching logic from thin wrapper pages directly
into their component files, eliminating the unnecessary indirection
where pages/X/index.tsx just imported and rendered components/X.tsx.

Consolidated pages:
- JudgesPage, LawyersPage, ClientsPage, MattersPage (navigation callbacks)
- ClientDetailPage, MatterDetailPage (data fetching + loading states)
- ClientMessagingPage (location.state extraction)
- PersonDetailPage (removed unused legacy component)

Each page is now self-contained with its own routing logic.
